Rework some logic in pause/resume
Prior to this, the following logic for pause/resume was:

1. When a request is paused, both EV_READ AND EV_WRITE are disabled.
2. When a request is resumed, if there is data waiting to be sent to the
   client, a "wait for the write to finish" flag is set, and ONLY EV_WRITE
   is enabled.
3. Once that part has been fully written, the "wait for write" flag is
   unset, and EV_READ is enabled again.

This can run into issues:

When EV_WRITE is enabled, and there is no more data that is written,
then nothing happens. This was fixed in libevent2.1x. with
bufferevent_trigger which is like event_active but you can trigger
specific data callbacks; on top of that it becomes deferred, which is
good in cases like this.

The new proposed logic is as follows:
1. Same
2. When a request is resumed, if data is awaiting, instead of
   just calling bufferevent_enable(), we also call event_active
   on the underlying bufferevent's write callback.
3. When a request is resumed and ready to read, a new event
   has been created: `resume_read_ev` which (in the next
   loop) will execute the readcb once it's activated.

Other:
Added debugging stuff everywhere if enabled.
Added a clearer example of pause/resume in examples/test_pause_resume.c